+++ jib-core/src/main/java/com/google/cloud/tools/jib/http/Connection.java
@@ -18,10 +18,12 @@ package com.google.cloud.tools.jib.http;
 
 import com.google.api.client.http.GenericUrl;
 import com.google.api.client.http.HttpMethods;
+import com.google.api.client.http.HttpRequest;
 import com.google.api.client.http.HttpRequestFactory;
 import com.google.api.client.http.HttpResponse;
 import com.google.api.client.http.apache.ApacheHttpTransport;
 import com.google.api.client.http.javanet.NetHttpTransport;
+import com.google.common.base.Preconditions;
 import java.io.Closeable;
 import java.io.IOException;
 import java.net.URL;
@@ -29,7 +31,8 @@ import javax.annotation.Nullable;
 import org.apache.http.NoHttpResponseException;
 
 /**
- * Sends an HTTP {@link Request} and stores the {@link Response}.
+ * Sends an HTTP {@link Request} and stores the {@link Response}. Clients should not send more than
+ * one request.
  *
  * <p>Example usage:
  *
@@ -117,11 +120,18 @@ public class Connection implements Closeable {
    * @throws IOException if building the HTTP request fails.
    */
   public Response send(String httpMethod, Request request) throws IOException {
-    httpResponse =
+    Preconditions.checkState(httpResponse == null, "Connection can send only one request");
+
+    HttpRequest httpRequest =
         requestFactory
             .buildRequest(httpMethod, url, request.getHttpContent())
-            .setHeaders(request.getHeaders())
-            .execute();
+            .setHeaders(request.getHeaders());
+    if (request.getHttpTimeout() != null) {
+      httpRequest.setConnectTimeout(request.getHttpTimeout());
+      httpRequest.setReadTimeout(request.getHttpTimeout());
+    }
+
+    httpResponse = httpRequest.execute();
     return new Response(httpResponse);
   }
 }